Move ChainedDataset to separate module
